requirements.txt
Flask==2.3.2
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
gunicorn
